requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
torch>=1.7.0
lpips>=0.1.4
av>=9.0.0
colorama>=0.4.4
scikit-image>=0.18.0 